+++ src/main/java/fr/eni/encheres/bo/Categorie.java
@@ -6,24 +6,24 @@ import java.util.Objects;
 public class Categorie implements Serializable {
 
     private static final long serialVersionUID = 1L;
-private int no_categorie;
+private Long no_categorie;
 private String libelle;
 
 public Categorie() {
 
 }
 
-public Categorie(int no_categorie, String libelle) {
+public Categorie(Long no_categorie, String libelle) {
     this.no_categorie = no_categorie;
     this.libelle = libelle;
 }
 
 
-    public int getNo_categorie() {
+    public Long getNo_categorie() {
         return no_categorie;
     }
 
-    public void setNo_categorie(int no_categorie) {
+    public void setNo_categorie(Long no_categorie) {
         this.no_categorie = no_categorie;
     }
